analysis/requirements.txt
aiocoap==0.4.5
anyio==3.7.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.99.1
fastapi-utils==0.2.1
h11==0.14.0
idna==3.4
influxdb-client==1.36.1
interchange==2021.0.4
kafka-python==2.0.2
monotonic==1.6
neo4j-driver==4.4.10
neobolt==1.7.17
packaging==23.1
paho-mqtt==1.6.1
pansi==2020.7.3
passlib==1.7.4
py2neo==2021.2.3
pyasn1==0.5.0
pybind11==2.10.4
pycountry==22.3.5
pydantic==1.10.10
Pygments==2.15.1
pymongo==4.4.0
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
reactivex==4.0.4
regex==2023.6.3
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.27.0
tld==0.13
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
validators==0.20.0
